chore: Extract string-related code to doctest/parts/string.h
