+++ CMakeLists.txt
@@ -67,6 +67,7 @@ if(${DOCTEST_WITH_MAIN_IN_STATIC_LIB})
         ${PROJECT_NAME}_with_main STATIC EXCLUDE_FROM_ALL
         ${doctest_parts_folder}/private/doctest.cpp
         ${doctest_parts_folder}/private/debugger.cpp
+        ${doctest_parts_folder}/private/string.cpp
     )
     add_library(${PROJECT_NAME}::${PROJECT_NAME}_with_main ALIAS ${PROJECT_NAME}_with_main)
     target_compile_definitions(${PROJECT_NAME}_with_main PRIVATE
